xray_fov/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asttokens==3.0.1
build==1.3.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
deflate_dict==1.2.2
defusedxml==0.7.1
dict_hash==1.3.7
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
ImageIO==2.37.2
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
monai==1.5.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
odfpy==1.4.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
pydicom==3.0.1
Pygments==2.19.2
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-calamine==0.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyxlsb==1.0.10
PyYAML==6.0.3
pyzmq==27.1.0
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
stack-data==0.6.3
stringzilla==4.2.3
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
xlrd==2.0.2
xlsxwriter==3.2.9
